Refactor: Move getTimeAgo function to dateFormatter.js and import it in TeamDetailFeedbackCards.jsx
